refactor: change struct name  and add comment, refactor some types
