+++ api/pkg/client/mlflow/mlflow.go
@@ -7,15 +7,6 @@ import (
 	"net/http"
 )
 
-type mlflowClient struct {
-	Client *http.Client
-	Config Config
-}
-
-type Config struct {
-	TrackingURL string
-}
-
 type Mlflow interface {
 	SearchRunForExperiment(idExperiment string) (SearchRunsResponse, error)
 	SearchRunData(idRun string) (SearchRunResponse, error)
@@ -23,67 +14,34 @@ type Mlflow interface {
 	DeleteRun(idRun string) error
 }
 
-type DeleteExperimentRequest struct {
-	ExperimentId string `json:"experiment_id" required:"true"`
-}
-
-type DeleteRunRequest struct {
-	RunId string `json:"run_id" required:"true"`
-}
-
-type DeleteExperimentErrorResponse struct {
-	ErrorCode string `json:"error_code"`
-	Message   string `json:"message"`
-}
-
-type SearchRunRequest struct {
-	ExperimentId []string `json:"experiment_ids" required:"true"`
-}
-
-type TagRun struct {
-	Key   string `json:"key"`
-	Value string `json:"value"`
-}
-type DataRun struct {
-	Tags []TagRun `json:"tags"`
-}
-
-type InfoRun struct {
-	RunId          string `json:"run_id"`
-	ExperimentId   string `json:"experiment_id"`
-	UserId         string `json:"user_id"`
-	LifecycleStage string `json:"lifecycle_stage"`
-	ArtifactURI    string `json:"artifact_uri"`
-}
-type RunResponse struct {
-	Info InfoRun `json:"info"`
-	Data DataRun `json:"data"`
-}
-type SearchRunsResponse struct {
-	RunsData []RunResponse `json:"runs"`
-}
-type SearchRunResponse struct {
-	RunData RunResponse `json:"run"`
+type mlflowClient struct {
+	Api    *http.Client
+	Config Config
 }
 
 func NewMlflowClient(httpClient *http.Client, config Config) *mlflowClient {
 	return &mlflowClient{
-		Client: httpClient,
+		Api:    httpClient,
 		Config: config,
 	}
 }
 
-func (mfc *mlflowClient) httpCall(method string, url string, headers map[string]string, body []byte, response interface{}) error {
+func (mfc *mlflowClient) httpCall(method string, url string, body []byte, response interface{}) error {
 	req, err := http.NewRequest(method, url, bytes.NewBuffer(body))
 	if err != nil {
 		return err
 	}
 
-	for key, value := range headers {
-		req.Header.Set(key, value)
+	if method == "POST" {
+		headers := map[string]string{
+			"Content-Type": "application/json",
+		}
+		for key, value := range headers {
+			req.Header.Set(key, value)
+		}
 	}
 
-	resp, err := mfc.Client.Do(req)
+	resp, err := mfc.Api.Do(req)
 	if err != nil {
 		return err
 	}
@@ -113,17 +71,13 @@ func (mfc *mlflowClient) SearchRunForExperiment(idExperiment string) (SearchRuns
 
 	searchRunURL := fmt.Sprintf("%s/api/2.0/mlflow/runs/search", mfc.Config.TrackingURL)
 
-	headers := map[string]string{
-		"Content-Type": "application/json",
-	}
-
 	input := SearchRunRequest{ExperimentId: []string{idExperiment}}
 	jsonInput, err := json.Marshal(input)
 	if err != nil {
 		return responseObject, err
 	}
 
-	err = mfc.httpCall("POST", searchRunURL, headers, jsonInput, &responseObject)
+	err = mfc.httpCall("POST", searchRunURL, jsonInput, &responseObject)
 	if err != nil {
 		return responseObject, err
 	}
@@ -136,7 +90,7 @@ func (mfc *mlflowClient) SearchRunData(idRun string) (SearchRunResponse, error) 
 	var runResponse SearchRunResponse
 	getRunURL := fmt.Sprintf("%s/api/2.0/mlflow/runs/get?run_id=%s", mfc.Config.TrackingURL, idRun)
 
-	err := mfc.httpCall("GET", getRunURL, nil, nil, &runResponse)
+	err := mfc.httpCall("GET", getRunURL, nil, &runResponse)
 	if err != nil {
 		return runResponse, err
 	}
@@ -149,16 +103,12 @@ func (mfc *mlflowClient) DeleteExperiment(idExperiment string) error {
 	// HIT Delete Experiment API
 	delExpURL := fmt.Sprintf("%s/api/2.0/mlflow/experiments/delete", mfc.Config.TrackingURL)
 
-	headers := map[string]string{
-		"Content-Type": "application/json",
-	}
-
 	jsonInput, err := json.Marshal(input)
 	if err != nil {
 		return err
 	}
 
-	err = mfc.httpCall("POST", delExpURL, headers, jsonInput, nil)
+	err = mfc.httpCall("POST", delExpURL, jsonInput, nil)
 	if err != nil {
 		return err
 	}
@@ -171,16 +121,12 @@ func (mfc *mlflowClient) DeleteRun(idRun string) error {
 	// HIT Delete Run API
 	delRunURL := fmt.Sprintf("%s/api/2.0/mlflow/runs/delete", mfc.Config.TrackingURL)
 
-	headers := map[string]string{
-		"Content-Type": "application/json",
-	}
-
 	jsonInput, err := json.Marshal(input)
 	if err != nil {
 		return err
 	}
 
-	err = mfc.httpCall("POST", delRunURL, headers, jsonInput, nil)
+	err = mfc.httpCall("POST", delRunURL, jsonInput, nil)
 	if err != nil {
 		return err
 	}
